refactor: 디렉토리명 수정
